fixed data type error in timezones table

+++ server/database/migrations/20200713100605_timezones.js
@@ -3,7 +3,7 @@ exports.up = function(knex) {
   return knex.schema.createTable('timezones', tbl => {
       tbl.increments()
       tbl.string('value')
-      tbl.integer('dif')
+      tbl.float('dif')
       tbl.string('name')
     
   })
